update render code, use instanced draw instead of loop

+++ src/text/color.h
@@ -4,7 +4,7 @@ namespace ftdgl {
 namespace text {
 
 struct color_s {
-    double r, g, b, a;
+    float r, g, b, a;
 };
 
 } //namespace text
