+++ src/text/text_buffer.h
@@ -11,9 +11,9 @@ namespace ftdgl {
 namespace text {
 
 typedef struct  __text_attr_s {
-    double bounds[4];
-    double color[4];
-    double back_color[4];
+    float bounds[4];
+    float color[4];
+    float back_color[4];
 } text_attr_s;
 
 class TextBuffer {
